requirements.txt
asgiref==3.7.2
certifi==2024.2.2
distlib==0.3.8
Django==5.0.3
django-colorfield==0.11.0
django-multiple-auth==1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.25.1
